refactor: consolidate network code
