+++ common/api/common.api
@@ -3,69 +3,6 @@ public final class io/opentelemetry/android/common/RumConstants {
 	public static final field OTEL_RUM_LOG_TAG Ljava/lang/String;
 }
 
-public final class io/opentelemetry/android/common/internal/features/networkattributes/CurrentNetworkAttributesExtractor {
-	public fun <init> ()V
-	public final fun extract (Lio/opentelemetry/android/common/internal/features/networkattributes/data/CurrentNetwork;)Lio/opentelemetry/api/common/Attributes;
-}
-
-public final class io/opentelemetry/android/common/internal/features/networkattributes/data/Carrier {
-	public fun <init> ()V
-	public fun <init> (I)V
-	public fun <init> (ILjava/lang/String;)V
-	public fun <init> (ILjava/lang/String;Ljava/lang/String;)V
-	public fun <init> (ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
-	public fun <init> (ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
-	public synthetic fun <init> (ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun component1 ()I
-	public final fun component2 ()Ljava/lang/String;
-	public final fun component3 ()Ljava/lang/String;
-	public final fun component4 ()Ljava/lang/String;
-	public final fun component5 ()Ljava/lang/String;
-	public final fun copy (ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)Lio/opentelemetry/android/common/internal/features/networkattributes/data/Carrier;
-	public static synthetic fun copy$default (Lio/opentelemetry/android/common/internal/features/networkattributes/data/Carrier;ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ILjava/lang/Object;)Lio/opentelemetry/android/common/internal/features/networkattributes/data/Carrier;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getId ()I
-	public final fun getIsoCountryCode ()Ljava/lang/String;
-	public final fun getMobileCountryCode ()Ljava/lang/String;
-	public final fun getMobileNetworkCode ()Ljava/lang/String;
-	public final fun getName ()Ljava/lang/String;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
-public final class io/opentelemetry/android/common/internal/features/networkattributes/data/CurrentNetwork {
-	public fun <init> (Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;)V
-	public fun <init> (Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;Lio/opentelemetry/android/common/internal/features/networkattributes/data/Carrier;)V
-	public fun <init> (Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;Lio/opentelemetry/android/common/internal/features/networkattributes/data/Carrier;Ljava/lang/String;)V
-	public synthetic fun <init> (Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;Lio/opentelemetry/android/common/internal/features/networkattributes/data/Carrier;Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun component1 ()Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public final fun component3 ()Ljava/lang/String;
-	public final fun copy (Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;Lio/opentelemetry/android/common/internal/features/networkattributes/data/Carrier;Ljava/lang/String;)Lio/opentelemetry/android/common/internal/features/networkattributes/data/CurrentNetwork;
-	public static synthetic fun copy$default (Lio/opentelemetry/android/common/internal/features/networkattributes/data/CurrentNetwork;Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;Lio/opentelemetry/android/common/internal/features/networkattributes/data/Carrier;Ljava/lang/String;ILjava/lang/Object;)Lio/opentelemetry/android/common/internal/features/networkattributes/data/CurrentNetwork;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getCarrierCountryCode ()Ljava/lang/String;
-	public final fun getCarrierIsoCountryCode ()Ljava/lang/String;
-	public final fun getCarrierName ()Ljava/lang/String;
-	public final fun getCarrierNetworkCode ()Ljava/lang/String;
-	public final fun getState ()Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public final fun getSubType ()Ljava/lang/String;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
-public final class io/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState : java/lang/Enum {
-	public static final field NO_NETWORK_AVAILABLE Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public static final field TRANSPORT_CELLULAR Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public static final field TRANSPORT_UNKNOWN Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public static final field TRANSPORT_VPN Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public static final field TRANSPORT_WIFI Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public static final field TRANSPORT_WIRED Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public static fun getEntries ()Lkotlin/enums/EnumEntries;
-	public final fun getHumanName ()Ljava/lang/String;
-	public static fun valueOf (Ljava/lang/String;)Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-	public static fun values ()[Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
-}
-
 public final class io/opentelemetry/android/common/internal/utils/AndroidUtilsKt {
 	public static final fun getThreadIdCompat (Ljava/lang/Thread;)J
 }
