+++ core/build.gradle.kts
@@ -61,6 +61,7 @@ dependencies {
 
     implementation(project(":agent-api"))
     implementation(project(":instrumentation:android-instrumentation"))
+    implementation(project(":instrumentation:network"))
     implementation(project(":common"))
     implementation(project(":semconv"))
     implementation(project(":services"))
