+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkApplicationListener.kt
@@ -5,10 +5,7 @@
 
 package io.opentelemetry.android.instrumentation.network
 
-import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork
 import io.opentelemetry.android.internal.services.applifecycle.ApplicationStateListener
-import io.opentelemetry.android.internal.services.network.CurrentNetworkProvider
-import io.opentelemetry.android.internal.services.network.NetworkChangeListener
 import io.opentelemetry.android.semconv.events.NetworkChangeEvent
 import io.opentelemetry.api.common.Attributes
 import io.opentelemetry.api.logs.Logger
