+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkChangeInstrumentation.kt
@@ -8,10 +8,8 @@ package io.opentelemetry.android.instrumentation.network
 import android.content.Context
 import com.google.auto.service.AutoService
 import io.opentelemetry.android.OpenTelemetryRum
-import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork
 import io.opentelemetry.android.instrumentation.AndroidInstrumentation
 import io.opentelemetry.android.internal.services.Services.Companion.get
-import io.opentelemetry.android.internal.services.network.NetworkProviderHolder
 import io.opentelemetry.api.common.AttributesBuilder
 
 /**
