+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkChangeListener.kt
@@ -3,9 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.internal.services.network
-
-import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork
+package io.opentelemetry.android.instrumentation.network
 
 /**
  * This class is internal and not for public use. Its APIs are unstable and can change at any time.
