+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkDetectorImpl.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.internal.services.network
+package io.opentelemetry.android.instrumentation.network
 
 import android.content.Context
 import android.net.ConnectivityManager
@@ -13,9 +13,7 @@ import android.telephony.TelephonyManager
 import android.util.Log
 import androidx.annotation.RequiresApi
 import io.opentelemetry.android.common.RumConstants
-import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork
-import io.opentelemetry.android.common.internal.features.networkattributes.data.NetworkState
-import io.opentelemetry.android.common.internal.features.networkattributes.data.NetworkState.TRANSPORT_CELLULAR
+import io.opentelemetry.android.instrumentation.network.NetworkState.TRANSPORT_CELLULAR
 
 /**
  * Implementation of NetworkDetector interface.
