+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkProviderHolder.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.internal.services.network
+package io.opentelemetry.android.instrumentation.network
 
 import android.annotation.SuppressLint
 import android.content.Context
