+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkUtils.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.internal.services.network
+package io.opentelemetry.android.instrumentation.network
 
 import android.Manifest
 import android.content.Context
